+++ routes/users.js
@@ -11,9 +11,6 @@ router.post(
   validateSchema(userValidationSchema),
   async (req, res) => {
     try {
-      console.log(req.body);
-      req.body.password = "citra2024";
-      console.log(req.body);
       const user = new User(req.body);
       await user.save();
       res.status(201).json({ message: "User registered successfully", user });
@@ -31,32 +28,23 @@ router.post("/login", async (req, res) => {
     if (!user || !(await bcrypt.compare(password, user.password))) {
       return res.status(401).json({ error: "Invalid email or password" });
     }
-    // Generate and return token (implementation depends on chosen strategy, e.g., JWT)
-    res.status(200).json({ message: "Login successful", user }); // Simplified for brevity
+    res.status(200).json({ message: "Login successful", user });
   } catch (error) {
     res.status(400).json({ error: error.message });
   }
 });
 
-// // Update User Info
-// router.patch('/update/:userId', async (req, res) => {
-//   const { userId } = req.params;
-//   try {
-//     const user = await User.findByIdAndUpdate(userId, req.body, { new: true });
-//     res.status(200).json(user);
-//   } catch (error) {
-//     res.status(400).json({ error: error.message });
-//   }
-// });
-
 // Get User Info
 router.get("/get-user/:userId", async (req, res) => {
   const { userId } = req.params;
-  console.log(userId);
+  const requesterId = req.user._id; // Assuming user ID is stored in req.user
   try {
     const user = await User.findById(userId);
-    if (!user) {
-      return res.status(404).json({ error: "User not found" });
+    if (
+      !user ||
+      (user._id.toString() !== requesterId && req.user.role !== "admin")
+    ) {
+      return res.status(404).json({ error: "User not found or access denied" });
     }
     res.status(200).json(user);
   } catch (error) {
@@ -64,9 +52,14 @@ router.get("/get-user/:userId", async (req, res) => {
   }
 });
 
-// Retrieve all users
+// Retrieve all users (Admin only)
 router.get("/all", async (req, res) => {
+  const requesterId = req.user._id; // Assuming user ID is stored in req.user
   try {
+    const requestingUser = await User.findById(requesterId);
+    if (requestingUser.role !== "admin") {
+      return res.status(403).json({ message: "Access denied" });
+    }
     const users = await User.find({});
     res.status(200).json(users);
   } catch (error) {
@@ -80,25 +73,34 @@ router.patch(
   validateSchema(userValidationSchema),
   async (req, res) => {
     const { userId } = req.params;
+    const requesterId = req.user._id; // Assuming user ID is stored in req.user
     try {
-      const user = await User.findByIdAndUpdate(userId, req.body, {
+      const user = await User.findById(requesterId);
+      if (userId !== requesterId && user.role !== "admin") {
+        return res.status(403).json({ message: "Access denied" });
+      }
+      const updatedUser = await User.findByIdAndUpdate(userId, req.body, {
         new: true,
       });
-      if (!user) {
+      if (!updatedUser) {
         return res.status(404).json({ message: "User not found" });
       }
-      res.status(200).json(user);
+      res.status(200).json(updatedUser);
     } catch (error) {
-      console.log(error);
       res.status(400).json({ error: error.message });
     }
   }
 );
 
-// Delete a user
+// Delete a user (Admin only)
 router.delete("/:userId", async (req, res) => {
   const { userId } = req.params;
+  const requesterId = req.user._id; // Assuming user ID is stored in req.user
   try {
+    const requestingUser = await User.findById(requesterId);
+    if (requestingUser.role !== "admin") {
+      return res.status(403).json({ message: "Access denied" });
+    }
     const deletedUser = await User.findByIdAndDelete(userId);
     if (!deletedUser) {
       return res.status(404).json({ message: "User not found" });
